Add special handling of "TASK_GRID" argument for predictions (#41)
In some cases the prediction API endpoint will recieve a special
argument, "TARGET_GRID", that is valid geojson, but we cannot pass in
as an argument to AWS Batch because it will increase the batch command
payload beyond the size restriction AWS imposes. In this case the API
will subsitute the argument for an S3 URI to where the geojson is
written to.